+++ src/services/data/documents.ts
@@ -52,10 +52,18 @@ export default (client, httpAuth: HttpInstance): DataDocumentsService => ({
    */
   async create<CustomData = null>(
     schemaId: ObjectId,
-    requestBody: Record<string, any>
+    requestBody: Record<string, any>,
+    options?: { gzip?: boolean }
   ): Promise<Document<CustomData>> {
-    return (await client.post(httpAuth, `/${schemaId}/documents`, requestBody))
-      .data;
+    return (
+      await client.post(
+        httpAuth,
+        `/${schemaId}/documents`,
+        requestBody,
+        {},
+        options
+      )
+    ).data;
   },
 
   /**
@@ -141,7 +149,7 @@ export default (client, httpAuth: HttpInstance): DataDocumentsService => ({
    * @param documentId The id of the targeted document.
    * @param rql Add filters to the requested list.
    * @param requestBody
-   * @returns any Success
+   * @returns AffectedRecords
    */
   async update(
     schemaId: ObjectId,
@@ -175,9 +183,9 @@ export default (client, httpAuth: HttpInstance): DataDocumentsService => ({
    * `DELETE_DOCUMENTS` | `global` | Delete the document
    * @param schemaId The id of the targeted schema.
    * @param documentId The id of the targeted document.
-   * @returns any Success
+   * @returns AffectedRecords
    */
-  async delete(
+  async remove(
     schemaId: ObjectId,
     documentId: ObjectId
   ): Promise<AffectedRecords> {
@@ -197,10 +205,10 @@ export default (client, httpAuth: HttpInstance): DataDocumentsService => ({
    * @param schemaId The id of the targeted schema.
    * @param documentId The id of the targeted document.
    * @param rql Add filters to the requested list.
-   * @param requestBody
-   * @returns any Success
+   * @param requestBody list of fields
+   * @returns AffectedRecords
    */
-  async deleteFields(
+  async removeFields(
     schemaId: ObjectId,
     documentId: ObjectId,
     requestBody: {
@@ -233,7 +241,7 @@ export default (client, httpAuth: HttpInstance): DataDocumentsService => ({
    * @param documentId The id of the targeted document.
    * @param rql Add filters to the requested list.
    * @param requestBody
-   * @returns any Success
+   * @returns AffectedRecords
    * @throws {IllegalArgumentError}
    * @throws {ResourceUnknownError}
    */
@@ -266,8 +274,8 @@ export default (client, httpAuth: HttpInstance): DataDocumentsService => ({
    * `UPDATE_ACCESS_TO_DOCUMENT` | `global` | **Required** for this endpoint
    * @param schemaId The id of the targeted schema.
    * @param documentId The id of the targeted document.
-   * @param requestBody
-   * @returns any Success
+   * @param requestBody list of groupIds
+   * @returns AffectedRecords
    */
   async linkGroups(
     schemaId: ObjectId,
@@ -298,8 +306,8 @@ export default (client, httpAuth: HttpInstance): DataDocumentsService => ({
    * `UPDATE_ACCESS_TO_DOCUMENT` | `global` | **Required** for this endpoint
    * @param schemaId The id of the targeted schema.
    * @param documentId The id of the targeted document.
-   * @param requestBody
-   * @returns any Success
+   * @param requestBody list of groupIds
+   * @returns AffectedRecords
    */
   async unlinkGroups(
     schemaId: ObjectId,
@@ -328,8 +336,8 @@ export default (client, httpAuth: HttpInstance): DataDocumentsService => ({
    * Note: When GroupSyncMode.LINKED_USERS_PATIENT_ENLISTMENT is set for a document, all the groups where the specified user is enlisted as patient will also be added to the document.
    * @param schemaId The id of the targeted schema.
    * @param documentId The id of the targeted document.
-   * @param requestBody
-   * @returns any Success
+   * @param requestBody list of userIds
+   * @returns AffectedRecords
    */
   async linkUsers(
     schemaId: ObjectId,
@@ -362,8 +370,8 @@ export default (client, httpAuth: HttpInstance): DataDocumentsService => ({
    * Note: When GroupSyncMode.LINKED_USERS_PATIENT_ENLISTMENT is set for a document, all the groups where the specified user is enlisted as patient will also be removed from the document.
    * @param schemaId The id of the targeted schema.
    * @param documentId The id of the targeted document.
-   * @param requestBody
-   * @returns any Success
+   * @param requestBody list of userIds
+   * @returns AffectedRecords
    */
   async unlinkUsers(
     schemaId: ObjectId,