+++ src/services/data/statuses.ts
@@ -9,8 +9,8 @@ export default (client, httpAuth: HttpInstance): DataStatusesService => ({
    * - | - | -
    * `UPDATE_SCHEMAS` | `global` | **Required** for this endpoint
    * @param schemaId The id of the targeted schema.
-   * @param requestBody
-   * @returns any Success
+   * @param requestBody The name and status data
+   * @returns AffectedRecords
    * @throws {ResourceAlreadyExistsError}
    */
   async create(
@@ -31,8 +31,8 @@ export default (client, httpAuth: HttpInstance): DataStatusesService => ({
    * `UPDATE_SCHEMAS` | `global` | **Required** for this endpoint
    * @param schemaId The id of the targeted schema.
    * @param name The name of the targeted status.
-   * @param requestBody
-   * @returns any Success
+   * @param requestBody The status data
+   * @returns AffectedRecords
    * @throws {ResourceUnknownError}
    */
   async update(
@@ -52,11 +52,11 @@ export default (client, httpAuth: HttpInstance): DataStatusesService => ({
    * `UPDATE_SCHEMAS` | `global` | **Required** for this endpoint
    * @param schemaId The id of the targeted schema.
    * @param name The name of the targeted status.
-   * @returns any Success
+   * @returns AffectedRecords
    * @throws {StatusInUseError}
    * @throws {ResourceUnknownError}
    */
-  async delete(schemaId: ObjectId, name: string): Promise<AffectedRecords> {
+  async remove(schemaId: ObjectId, name: string): Promise<AffectedRecords> {
     return (await client.delete(httpAuth, `/${schemaId}/statuses/${name}`))
       .data;
   },