+++ src/services/users/groupRoles.ts
@@ -21,7 +21,7 @@ export default (
    * - | - | -
    * none |  | Everyone can use this endpoint
    *
-   * @returns any Success
+   * @returns PagedResult<GlobalPermission>
    */
   async getPermissions(): Promise<PagedResult<GlobalPermission>> {
     return (await userClient.get(httpWithAuth, '/groups/permissions')).data;
@@ -36,7 +36,7 @@ export default (
    *
    * @param groupId Id of the targeted group
    * @param rql Add filters to the requested list.
-   * @returns any Success
+   * @returns PagedResult<GroupRole>
    */
   async get(
     groupId: ObjectId,
@@ -60,8 +60,8 @@ export default (
    * `CREATE_GROUP_ROLE` | `global` | Create a role for the group
    *
    * @param groupId Id of the targeted group
-   * @param requestBody
-   * @returns any Success
+   * @param requestBody The role to add
+   * @returns GroupRole
    */
   async add(groupId: ObjectId, requestBody: AddRole): Promise<GroupRole> {
     return (
@@ -82,8 +82,8 @@ export default (
    *
    * @param groupId Id of the targeted group
    * @param roleId Id of the targeted role
-   * @param requestBody
-   * @returns any Success
+   * @param requestBody The role data to update
+   * @returns GroupRole
    * @throws {ResourceUnknownError}
    */
   async update(