+++ tests/unit/data/documentsService.test.ts
@@ -93,7 +93,7 @@ describe('Documents Service', () => {
     nock(`${host}${DATA_BASE}`)
       .delete(`/${schemaId}/documents/${documentId}`)
       .reply(200, { affectedRecords: 1 });
-    const res = await sdk.data.documents.delete(schemaId, documentId);
+    const res = await sdk.data.documents.remove(schemaId, documentId);
     expect(res.affectedRecords).toBe(1);
   });
 
@@ -101,7 +101,7 @@ describe('Documents Service', () => {
     nock(`${host}${DATA_BASE}`)
       .post(`/${schemaId}/documents/${documentId}/deleteFields`)
       .reply(200, { affectedRecords: 1 });
-    const res = await sdk.data.documents.deleteFields(schemaId, documentId, {
+    const res = await sdk.data.documents.removeFields(schemaId, documentId, {
       fields: ['location.latitude'],
     });
     expect(res.affectedRecords).toBe(1);